Refactor coordinate transformation functions and enhance testing for consistency
- Introduced new functions _substitute_coordinates, solve_real, and solve_complex in coordinates.py to streamline coordinate transformations.
- Updated existing transformation logic to utilize the new functions, improving clarity and maintainability.
- Refactored test cases in test_coordinates.py to align with the new transformation functions, ensuring consistency in coordinate transformations.
- Removed deprecated test functions and improved test coverage for coordinate transformations in the testing suite.
